backend/requirements.txt
asgiref
certifi
charset-normalizer
Django
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
idna
Markdown
Pillow
psycopg2-binary
PyJWT
python-decouple
pytz
requests
sqlparse
urllib3
